fix f32 compatibility

+++ crates/ferrum_physics/src/update.rs
@@ -12,10 +12,8 @@ pub struct Physics {
 }
 
 impl Physics{
-    pub fn physics_update(&mut self, dt: &mut Float) {
-
-
-        let dt = *dt;
+    pub fn physics_update(&mut self, dt: &mut f64) {
+        let dt = *dt as Float;
         // Mutably access an instance and change its position
         let snapshot = self.rigidbodies.clone();
         let mut next_bodies: RigidBodySet = self.rigidbodies.clone();
